refactor scattergl - add plot and style files
